+++ HOPSAN++/HopsanGUI/GUIPort.cpp
@@ -177,7 +177,7 @@ void GUIPort::plot(size_t nVar) //En del vansinne i denna metoden...
     QVector<double> time = QVector<double>::fromStdVector(*(this->mpCorePort->getTimeVectorPtr())); //Inte lampligt att skyffla data pa detta viset
     QVector<double> y(dataLength);// = QVector<double>::fromStdVector((this->mpCorePort->getDataVectorPtr()->at(1)));
     
-    for (int i = 0; i<dataLength; ++i) //Denna loop ar inte klok
+    for (size_t i = 0; i<dataLength; ++i) //Denna loop ar inte klok
     {
         //timeq[i] = this->mpCorePort->getTimeVectorPtr()->at(i);
         y[i] = (this->mpCorePort->getDataVectorPtr()->at(i)).at(nVar);
